refactor(api): fold resolve and check into routes/components.py
Everything under /components lives in one module on a single router —
no sub-router package.

By Digitl
